Find a word from file
The program prints File not found, if the given file cannot be found. If the file can be found but the user gives no word to search for (i.e. the word is empty or its length is zero), the program prints File found. If the file can be found and if the given word is non-empty, the program prints one of the two remaining choices according to the fact, whether the word occurs in the file or not.

You can use the Match case checkbox to define whether or not capitalization matters. Therefore, this choice has also effect on which of the two last choices the program prints. You need not store the made choice into a boolean variable, since the class QCheckBox has the function isChecked, which returns the made choice.

You can close the MainWindow with the Close button.